Download savu installer only if it is not already there
This makes testing a bit easier

+++ docker-image/get-and-install-savu.sh
@@ -17,7 +17,9 @@ if [[ "$SAVU_URL" == "" ]]; then
     exit -1
 fi
 
-wget -O savu.tar.gz "$SAVU_URL"
+if [[ ! -f "savu.tar.gz" ]]; then
+    wget -O savu.tar.gz "$SAVU_URL"
+fi
 
 tar -xvzf savu.tar.gz
